projekt/requirements.txt
google_api_python_client==2.160.0
google_auth_oauthlib==1.2.1
matplotlib==3.6.3
protobuf==5.29.3
PySide6==6.8.1.1
PySide6==6.8.1.1
PySide6_Addons==6.8.1.1
PySide6_Essentials==6.8.1.1
